+++ fdb-record-layer-lucene/src/test/java/com/apple/foundationdb/record/lucene/LuceneIndexTestValidator.java
@@ -117,6 +117,10 @@ public class LuceneIndexTestValidator {
     void validate(Index index, final Map<Tuple, Map<Tuple, Tuple>> expectedDocumentInformation,
                   final int repartitionCount, final String universalSearch, final boolean allowDuplicatePrimaryKeys) throws IOException {
         final int partitionHighWatermark = Integer.parseInt(index.getOption(LuceneIndexOptions.INDEX_PARTITION_HIGH_WATERMARK));
+        String partitionLowWaterMarkStr = index.getOption(LuceneIndexOptions.INDEX_PARTITION_LOW_WATERMARK);
+        final int partitionLowWatermark = partitionLowWaterMarkStr == null ?
+                                          LucenePartitioner.DEFAULT_PARTITION_LOW_WATERMARK :
+                                          Integer.parseInt(partitionLowWaterMarkStr);
         // If there is less than repartitionCount of free space in the older partition, we'll create a new partition
         // rather than moving fewer than repartitionCount
         int maxPerPartition = partitionHighWatermark;
@@ -156,28 +160,8 @@ public class LuceneIndexTestValidator {
                                 Tuple.fromBytes(partitionInfo.getTo().toByteArray()));
                     }
 
-                    int minPerPartition;
-                    if (partitionInfos.size() == 1) {
-                        // if there is only one partition, it should have exactly the number of documents, which is
-                        // verified below
-                        minPerPartition = 1;
-                    } else if (i == 0) {
-                        // if it is the oldest, it could have fewer if the test inserted max into the most recent, and
-                        // then inserted a couple that were older
-                        // If we add tests that don't try at all to order the timestamps this could become more
-                        // complicated
-                        minPerPartition = 1;
-                    } else if (i == partitionInfos.size() - 2) {
-                        // The second to last should have at least the repartitionCount that would have been moved out
-                        // of the most recent
-                        minPerPartition = Math.min(repartitionCount, partitionHighWatermark);
-                    } else {
-                        // Everything else should be filled as much as it can, but may have had repartitionCount moved
-                        // out.
-                        minPerPartition = Math.max(1, partitionHighWatermark - repartitionCount);
-                    }
                     assertThat("Group: " + groupingKey + " - " + allCounts, partitionInfo.getCount(),
-                            Matchers.allOf(lessThanOrEqualTo(maxPerPartition), greaterThanOrEqualTo(minPerPartition)));
+                            Matchers.allOf(lessThanOrEqualTo(maxPerPartition), greaterThanOrEqualTo(partitionLowWatermark)));
                     assertTrue(usedPartitionIds.add(partitionInfo.getId()), () -> "Duplicate id: " + partitionInfo);
                     final Tuple fromTuple = Tuple.fromBytes(partitionInfo.getFrom().toByteArray());
                     if (i > 0) {